Raise CALL-ARGUMENTS-LIMIT: APPLY spreads beyond the old 64-arg cap
OP_APPLY flattened its arglist into a fixed CL_Obj[64] thread buffer and
errored with "too many arguments (>64)" past that — so (apply #'f big-list)
failed for any list over 64 elements.  serapeum's POSITIVE/NEGATIVE-NUMERIC-
STRING+ tests apply #'string+ (a &rest function) to lists up to 1000 long and
tripped this.

Rework OP_APPLY to spread the arglist directly onto the (GC-rooted) VM stack
with no fixed buffer, bounded by the new CL_CALL_ARGS_LIMIT (4096) and the
stack size.  The &rest list is now consed straight off the stack slots rather
than through the 256-entry vm_extra_args buffer, lifting that cap too; cl_cons
protects its args and the stack slots are rooted, so the partial list survives
compaction (callee_bc is re-derived after the cons loop, as before).  Builtins
get the args as a rooted stack slice (an int-nargs ABI, no 255 cap) — this also
removes the latent bug where the old vm_flat_args buffer holding a builtin's
args was never GC-marked.

Widen CL_Frame.nargs to uint16_t so OP_ARGC reports correct counts past 255 for
&rest callees.  CALL-ARGUMENTS-LIMIT is now 4096 (independent of LAMBDA-
PARAMETERS-LIMIT, which still caps *declared* params at 64; a &rest function
accepts arbitrarily many actual args).

cl_vm_apply (the MAPCAR/REDUCE/JIT trampoline) no longer silently truncates at
255: builtins go up to CALL-ARGUMENTS-LIMIT, and the bytecode/closure stub path
(a one-byte OP_CALL operand) now errors clearly past 255 instead of dropping
args.  The JIT apply trampoline keeps a small C-stack buffer (255) to avoid
blowing the Amiga 64K stack; code needing more runs through the VM's inline
OP_APPLY.

No FASL bump: bytecode opcodes and operand encoding are unchanged; only the
runtime calling convention and the constant's value changed.

Tests: test_vm apply-over-64 (&rest/required+rest/builtin/ordered-sum, up to
1000) + call-arguments-limit invariant; gc-stress APPLY-large-&rest case;
Amiga run-tests apply cases up to 255 (JIT path).  serapeum 7->5; host fast
tier + gc-stress 221 + Amiga 3410 all green.

+++ src/jit/jit.c
@@ -3436,7 +3436,7 @@ CL_Obj cl_jit_invoke(CL_Obj func_obj, CL_Bytecode *bc, int nargs)
         sf->ip        = 0;
         sf->bp        = (uint32_t)(cl_vm.sp - nargs);
         sf->n_locals  = nargs;
-        sf->nargs     = (uint8_t)nargs;
+        sf->nargs     = (uint16_t)nargs;
         sf->nlx_level = cl_nlx_top;
         pushed_frame  = 1;
     }
